aider: Update the jest timeout to be 60s.
# Aider chat conversation:

USER: update the jest timeout to be 60s

ASSISTANT: To update the jest timeout to be 60 seconds, you need to modify the jest.config.js file.

Here is the change:

```javascript
jest.config.js
<<<<<<< ORIGINAL
module.exports = {
  preset: 'ts-jest',
  testEnvironment: 'node',
};
=======
module.exports = {
  preset: 'ts-jest',
  testEnvironment: 'node',
  testTimeout: 60000,
};
>>>>>>> UPDATED
```

This change adds the `testTimeout` property to the Jest configuration, setting it to 60000 milliseconds (60 seconds).