aider: Update the /files endpoint to return a "no files" message if the response is empty.
# Aider chat conversation:

USER: update the /files endpoint to return a "no files" message if the response is empty.

ASSISTANT: To update the `/files` endpoint to return a "no files" message if the response is empty, we need to modify the `src/index.ts` file.

Here are the changes:

**src/index.ts**
```typescript
src/index.ts
<<<<<<< ORIGINAL
app.get("/files", async (req, res) => {
  try {
    const files = await listFilesInS3();
    let html = '<table>';
    files.forEach(file => {
      html += `<tr><td>${file}</td></tr>`;
    });
    html += '</table>';
    res.send(html);
  } catch (error) {
    res.send("ERROR");
  }
});
=======
app.get("/files", async (req, res) => {
  try {
    const files = await listFilesInS3();
    if (files.length === 0) {
      res.send("No files");
    } else {
      let html = '<table>';
      files.forEach(file => {
        html += `<tr><td>${file}</td></tr>`;
      });
      html += '</table>';
      res.send(html);
    }
  } catch (error) {
    res.send("ERROR");
  }
});
>>>>>>> UPDATED
```

This change will check if the `files` array is empty. If it is, it will send a "No files" message. Otherwise, it will create an HTML table with the filenames and send that.